requirements.txt
PyQt6>=6.0.0
requests>=2.28.0
atlassian-python-api>=3.41.0 